Moved API to utils.py
